refactor(renderer): settings store 拆出 provider-login store，登录取消时机从 setOpen 挪到 LoginDialog 卸载 cleanup
